Rename getPackageFiles.ts -> lfxApiClient.ts
